Remove todo: this test is not possible for huffman decoding.
The huffman decoder should never write past the output buffer.
This is only possible for RLE and LZSS, where the compressed stream
may contain 'instructions' that extend past the uncompressed size.

For the huffman decoder this is not possible. It simply decodes bit after
bit until it has decoded all symbols or prematurely reached the end
of the input bitstream.