Break test file.
Manually edited file, so that the bitstream is not aligned to a multiple
of 4 bytes anymore, but only two bytes. This requires not only moving
around the bitstream and truncating the file, but also fixing up the
tree size byte.

The decoder is not yet fixed, so this causes tests to fail.