g++: enable -Wunsafe-loop-optimizations
Not sure I understand this correctly: this warning will only be issued
if we use -funsafe-loop-optimizations, and the compiler actually does
perform an 'unsafe loop optimization'. If we don't use this option,
then enabling this warning should have no effect and thus cannot hurt.
If we do enable -funsafe-loop-optimizations (which we certainly won't
for the time being), then having -Wunsafe-loop-optimizations does
probably not hurt either. Anyway: we can always disable warnings if
they are annoying.