Add -Wpointer-arith: sounds useful.
This warns about arithmethic with function pointers and void pointers
which work in gcc but are really not defined. As I understand, this
warns us if we do things with gcc (g++?) that are non-standard and may
not work with other compilers, so that's a warning I'd like to have.
[no ci]