Fix RLE encoded reference data.
Data was created with CUE's RLE encoder, but that one fails to pad the
compressed data size to a multiple of 4 bytes. GBATEK states that the
size should be padded, and GBACrusher indeed does the padding, so we'll
assume that padding is correct. The new version of the file has two
0 padding bytes appended at its end, so that its size os 8 rather than 6
bytes.
[no ci]