Remove todo: for the time being we live with a TByteWriter type arg.
In principle we'd probably want to have a byte writer concept, but
I'll probably not do that for the time being. The reason we could
not use std::output_iterator was simple: we tried to define
writer as byte_writer<OutputIterator>& but the calling code was passing
an unbounded_byte_writer, which is simply not the same.
[no ci]