Remove todo items
We have our public interface, so that's taken care of.
We don't need to do a naive implementation anymore.
We went straigt to breadth first tree traversal, which is hopefully
what we need. And finally we've basically implemented it, we just
need to iron out a few things and do some more testing
[no ci]