Add internal_error exception.
It is deliberately not derived from agbpack_exception, but directly
from std::logic_error. We DO want this to be a logic error. The only
reason it is derived is because we want it to add the
"agbpack: internal error: " message prefix. We could just aswell have
created a creation function.
[no ci]