Remove pointless and confusing assertion.
That assertion does not test anything test worthy. Basically it just
tests whether the calculation right in front of it 2*x+1 worked correctly.
Moreover, the assertion itself is wrong/confusing, since it asserts
that the minimum size of the serialized tree is 2 bytes, but that's
what it can theoretically be. In practice it needs to be aligned at a
four byte boundary, so the minimum size is actually 4 bytes.

We do however not need another runtime check for this, since we already
have a check whether the bitstream is word-aligned
[no ci]