Delete old huffman_encoder test.
That section has been transformed into the new test using a subdirectory,
and not using reference data since using reference data is pointless.
Irt is pointless because different encoders can produce different output
that is still valid (so can be decoded by the same decoder) and is still
the smallest possible encoding.
[no ci]