New huffman_encoder_test: use huffman.good.8.0-bytes.txt
Since this file is used by the encoder test only we can move it to the
new subdirectory. This breaks the old zzz_test, but since it is now
a duplicate of the new huffman_encoder test we can delete its code.
There's still a lot of TODOs we need to go through, so keeping these.
[no ci]