Remove todo.
Using queue.size() rather than next to calculate the offset works because
the offset is relative. If we used next (which is absolute), then we
would also have to have a current variable and do something like
'next - current'. But we don't have to, because we can use queue.size()
[no ci]