Final review of huffman_tree_node
Reformatted the constructor for nodes.
Decided to keep all the deleted special members. They are not needed,
since they are implicitly deleted because we have a unique_ptr field.
If we do keep them however, then at least MSVC is able to print much
better error messages, so we keep them, for the time being anway.
Removed the argument names from the special members tho, since they
are really not needed
[skip ci]