Remove todo item regarding possible huffman encoder bug
The described case does indeed create an *encoded* huffman tree which
contains the number 0x41 (what we perceived as 'A') twice. However, in one
case it is the actual leaf node containing the symbol 'A', in the other
case it is an internal node which coincidentally happens to have the
value 0x41.

Things become clear if one uses different symbol values but same symbol
frequencies, e.g. all lower case (abbccddeeee instead of ABBCCDDEEEE)
[skip ci]