Remove todo regarding width of type symbol.
Testing this is somewhat pointless. A better way for the time being
is to have a static_assert on type symbol's maximum value with an
explanation where the requirement comes from. We *could* change our
code to work with an 8 bit wide symbol type, but why bother, this
is going tu run mostly on modern 64 bit CPUs.