Tick off -Wconversion: we don't want this.
Enabling this triggers a lot of warnings in rather neat and well-tested
code.

gcc wiki (https://gcc.gnu.org/wiki/NewWconversion):
"[...]  Wconversion is designed for a niche of uses (security audits,
porting 32 bit code to 64 bit, etc.) where the programmer is willing
to accept and workaround invalid warnings. Therefore, it shouldn't
be enabled if it is not explicitly requested."

OK, I'll believe it
[skip ci]