MSVC: tick off /permissive-: we don't need to mess around with this
From https://learn.microsoft.com/en-us/cpp/build/reference/permissive-standards-conformance:

"The /permissive- option is implicitly set by the /std:c++latest option
starting in Visual Studio 2019 version 16.8, and in version 16.11 by
the /std:c++20 option. /permissive- is required for C++20 Modules support."

Well we clearly are compiling a module, and we are using /std:c++latest
because (at the time of writing this commit message) we are using
set(CMAKE_CXX_STANDARD 23), so we must be already using /permissive-
and probably should not mess around with this and instead let CMake
do its job.
[skip ci]