Add static_assert that fails if size_t is too small
The way code is written at the moment it will probably not work when
compiled for a 16 bit target. It might partially work, but probably
it will not be able to handle the maximum size of uncompressed data
allowed by the GBA BIOS' compression methods.

Wanting to produce assets for a 32 bit target on a 16 bit target
is probably silly anyway, so we don't really need to support this.
But if we know it is not going to work we can just as well document this,
and we're using a static_assert to do so.