Replace CC0 by delta.cppm as test data file for lzss_encoder_test.
Rationale: reuse did not properly detect the copyright text of the
CC0-1.0 license. Decided to use one of our own source files where
copyright is clear and properly detected. It does not matter much,
the file should just be a larger piece of text and when lazy or
optimal LZSS parsing is used it should be smaller than when greedy
parsing is used, which is the case too.
[skip ci]