Remove todo: verified this
It is indeed so that the input 'aaa' is encoded as 3 literal bytes
by e.g. CUE lzss. Obviously it has to be the case: the first 'a'
cannot be a match because it is at the beginning of the file.
At the second 'a' we have a match of 2 bytes, but the minimum
encodable match length is 3 bytes (because we store count=0 and
the decoder adds 3), so the last two 'a' can only be literals too