Fix bug: array subscript out of range.
Basically we want to get an end iterator, but if the index is past
the last element, then &array[x] already bombs in debug mode when
the indexing is done, even though we don't read anything but only
take the address. Solution: index the last element, and increment
the pointer afterwards.

This gets us a step further. We start now seeing differences caused by
a) We still ignore VRAM safety
b) Probably we produce output different from the greedy encoder
[skip ci]