add_header: take string, not optional_string.
If we take optional_string and somebody actually does not supply
a string, we end up most likely with something useless/unintended.
For instance, if group is left at its default, that is, 0, then
calling add_header will actually terminate the list of options.
Most likely this is not what's intended.