Install clownlzss.h together with .cppm files.
This is needed because there is a .cppm files that includes clownlzss.h.
.cppm files in turn are compiled when the installed library is consumed
by another target, so private headers need to be installed. They way we
do this may be a workaround, respectively I'm not even sure this is
a use case covered by CMake ("install private headers such that they are
visible only when BMI's are built vor a consumed installed module")

Either way, it works so far.